Deactivate endpoint done, fixed URIs to make them REST compliant, next is adding new supplier to existing product

+++ src/main/java/com/eduardo/store/controller/ProductController.java
@@ -5,13 +5,14 @@ import com.eduardo.store.service.IProductService;
 import com.eduardo.store.service.ISupplierService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
 
 @CrossOrigin(origins = "http://localhost:8082")
 @RestController
-@RequestMapping("/api/product")
+@RequestMapping("/products")
 public class ProductController {
 
     @Autowired
@@ -20,48 +21,58 @@ public class ProductController {
     @Autowired
     private ISupplierService supplierService;
 
-    @GetMapping("/findAll")
+    @GetMapping
     @ResponseBody
-    public List<ProductDTO> findAll(){
-        return productService.findAll();
-    }
-
-    @GetMapping("/itemCode/{productItemCode}")
-    @ResponseBody
-    public List<ProductDTO> findByItemCode(@PathVariable Long productItemCode){
-        return productService.findByItemCode(productItemCode);
-    }
-
-    @GetMapping("/state/{productState}")
-    @ResponseBody
-    public List<ProductDTO> findByState(@PathVariable String productState){
-
-        return productService.findByState(productState);
+    public ResponseEntity<List<ProductDTO>> getByFilter(
+            @RequestParam(value = "item-code", required = false) Long productItemCode,
+            @RequestParam(value = "state", required = false) String productState) {
+        if (!(productItemCode == null || productService.findByItemCode(productItemCode).isEmpty())) {
+            return ResponseEntity.ok(productService.findByItemCode(productItemCode));
+        } else if (!(productState == null || productService.findByState(productState).isEmpty())) {
+            return ResponseEntity.ok(productService.findByState(productState));
+        } else if (!productService.findAll().isEmpty()) {
+            return ResponseEntity.ok(productService.findAll());
+        }
+        return ResponseEntity.badRequest().build();
     }
 
     @PostMapping
     @ResponseStatus(HttpStatus.CREATED)
     @ResponseBody
-    public ProductDTO save(@RequestBody ProductDTO product){
-        return productService.save(product);
+    public ResponseEntity<ProductDTO> create(@RequestBody ProductDTO product){
+        if (product == null) {
+            return ResponseEntity.badRequest().build();
+        }
+        return ResponseEntity.ok(productService.save(product));
     }
 
     @PutMapping("/{itemCode}")
     @ResponseBody
-    public ProductDTO update(@RequestBody ProductDTO productDTO,
+    public ResponseEntity<ProductDTO> update(@RequestBody ProductDTO productDTO,
                           @PathVariable Long itemCode) {
-        return productService.update(productDTO, itemCode);
+        if (productDTO == null || itemCode == null) {
+            return ResponseEntity.badRequest().build();
+        }
+        return ResponseEntity.ok(productService.update(productDTO, itemCode));
     }
 
     @DeleteMapping("/{itemCode}")
-    public void delete(@PathVariable Long itemCode){
+    @ResponseBody
+    public ResponseEntity<Void> delete(@PathVariable Long itemCode){
+        if (itemCode == null) {
+            return ResponseEntity.badRequest().build();
+        }
         productService.deleteByItemCode(itemCode);
+        return ResponseEntity.ok().build();
     }
 
-    @PutMapping({"/deactivate/{itemCode}"})
+    @PatchMapping({"/{itemCode}"})
     @ResponseBody
-    public ProductDTO deactivate(@PathVariable Long itemCode) {
-        return productService.deactivate(itemCode);
+    public ResponseEntity<ProductDTO> updateState(@PathVariable Long itemCode) {
+        if (itemCode == null) {
+            return ResponseEntity.badRequest().build();
+        }
+        return ResponseEntity.ok(productService.deactivate(itemCode));
     }
 }
